Week 4 Rollups

+++ nft_staking_q3/programs/nft_staking_q3/src/state/user_account.rs
@@ -3,7 +3,7 @@ use anchor_lang::prelude::*;
 #[account]
 #[derive(InitSpace)]
 pub struct UserAccount {
-    pub points: u32,
+    pub points: i64,
     pub amount_staked: u8,
     pub bump: u8,
 }
